Update dependencies to fix wheel build
Pip was failing to build the wheel for Prophet. To fix this, we explicitly include a compatible set of dependencies in requirements.txt, which are equivalent to the Prophet v0.6.0 requirements. We first install these with pip3, so that they are available to Prophet at build time (to say, we move Prophet installation to reflect this). The README has been updated to reflect the extra step needed if installing manually, and also to assume a CML/CDSW environment.